requirements.txt
streamlit
langchain
langchain-community
langchain-openai
pypdf
python-docx
tiktoken
faiss-cpu
numpy
pandas
python-dotenv
